Void PMC, w/ semantics like id as right-op and zero as left-
New Void PMC, which has semantics like an identity function when used
as a right operand and semantics like zero when used as a left operand
in arithmetic operations. Void converts to an empty string for
concatenation but is not convertible directly to a number because its
numeric value depends on its context.

In list context, Void contributes no elements to a list, so ',' is an
empty list and '5,', '<< 5', and '5 >>' are three different ways to
write a list of one element, the number 5. Void evaluates to Void when
indexed with «[ ]» and indexing any list with Void results in the list
itself.

Lastly, void is a no-op for sequencing «;». Void has the effect of
allowing trailing operators without changing the value of an
expression.