conjunction «&&» and disjunction «||» ops; int/float Vals
Add two new operators, conjunction «&&» and disjunction «||»,
currently implemented used :pasttype for short-circuiting. While
testing these operators, the following error was obtained when the
right operand was a numeric constant:

  Method 'arity' not found for invocant of class 'Float'

This error was resolved by modifying the integer and float actions
methods to make PAST::Val nodes of the appropriate type (Integer or
Float), rather than simplying converting the string to a number and
making that.