alias all the infix op func names to their short forms
For each of the infix operator functions, such as '&infix:<+>' and
'&infix:<*>', assign the function to an HLL-global variable named for
the operator. For example, store '&infix:<*>' in the HLL-global
variable '*'. This allows these operators to be called in the same way
as regular functions. For example:

  «+» 5 10;
  «<» 10 20;

Note that «,» currently has different semantics since it is
list-associative rather than left- or right-assocative and hence its
corresponding function defines one parameter, which is :slurpy -- the
function takes all its parameters at once but requires none. The
following example creates a list of one element, which is the only
thing «,» can currently be used for:

  «,» 10;