«' '» string can be an invocant because no interpolation
A single-quoted string constants (' ') can be an invocant because
interpolation does not occur in single-quoted strings. Double-quoted
string constants can not currently be invocants because with
interpolation they can contain variable declarations, which can be
broken by backtracking out of a variable originally parsed under
invocant in order to match a factor.