remove invocant rule; fix (x = 1), "\{x = 1}" lexical bug
Remove the invocant proto rule. A function_call is now an invocant
factor followed by zero or more factors, the arguments (i.e. the rule
now matches with zero arguments). A term is now always a function_call
(the factor instance of the term proto was removed), which may or may
not represent an actual function call, depending on whether or not any
arguments are present.

These changes fix the lexical-not-found bugs encounted for expressions
like (x = 1) and "\{x = 1}", which were a result of backtracking after
x was initially parsed as an invocant before the function_call rule
failed because '=' was encountered instead of a factor. Backtracking
left the @?BLOCK stack inconsistent with the AST.