function params are required by default; add «?» if optional
Removed the 'viviself' attribute from function parameters by default,
which makes the parameters required. The «?» unary postfix operator
may be added to a parameter to signal that it is optional.

The «!» unary function call operator will fail with a dispatch error
when invoked on a function with a required parameter.